+++ .flake8
@@ -1,4 +1,4 @@
 [flake8]
 max-line-length = 160
 exclude = .git,.tox,dist,deb_dist,__pycache__,._*
-ignore = E121,E123,E126,E226,E24,E704,W503,W504,W606
+ignore = E121,E123,E126,E226,E24,E704,W503,W504
